* 🎉 feat(httpbody.proto): add HttpBody message This commit adds the HttpBody message to the google.api package. The HttpBody message represents an arbitrary HTTP body and can be used for payload formats that can't be represented as JSON, such as raw binary or an HTML page. It can be used both in streaming and non-streaming API methods in the request as well as the response. The HttpBody message has three fields: content_type, data, and extensions. The content_type field specifies the HTTP Content-Type header value specifying the content type of the body. The data field contains the HTTP request/response body as raw binary. The extensions field is application-specific response metadata and must be set in the first response for streaming APIs.
* 🆕 feat(helpers.go): add MapToAmortizations function
This commit adds a new function called MapToAmortizations to the helpers.go file. This function maps the output of the CalcMortgageAmortization function to a protobuf MortgageAmortizationResponse message. The function takes in several parameters, including capital, terms, interestType, amortizationAmount, amortizationYear, and amortizationMonth, and returns a pointer to a MortgageAmortizationResponse message.

* 🚀 chore(nif.proto): add summary to GetAmortization operation
* 🚀 chore(nif.swagger.json): add summary to GetAmortization operation
* ✨ feat(nif_grpc.pb.go): add comment for Mortgage section
The GetAmortization operation now has a summary added to it, which improves the readability of the API documentation. The comment for the Mortgage section has been added to the nif_grpc.pb.go file, which provides context for the new operation.